Add specification of category blocks by config.xml
Makes the extension more extendible, to allow custom blocks to be ajax
layered by the frontend.

NOTE: This doesn't update the JS so you need to do that yourself in your
site if you want it to work. We need to overhaul the JS anyway. The
extension continues to work as shipped, ajax loading the layered nav and
the product list blocks onto the page.

Additionally, we apply a few small fixes to make the whole thing more
defensive, and in one case deleted a for loop that was never doing
anything.